+++ include/mator/4Dim/Vec4.h
@@ -11,16 +11,16 @@ namespace mtr
 		
 	public:
 		//components
-		float x;
-		float y;
-		float z;
-		float w;
+		double x;
+		double y;
+		double z;
+		double w;
 
 		//ctrs
 		Vec4();
-		Vec4(float nx, float ny, float nz, float nw);
-		Vec4(float mag, const Vec4 &dir);
-		Vec4(const Vec3 &vec, float nw=1);
+		Vec4(double nx, double ny, double nz, double nw);
+		Vec4(double mag, const Vec4 &dir);
+		Vec4(const Vec3 &vec, double nw=1);
 		Vec4(const Vec4 &vec);
 
 		//dtr
@@ -30,28 +30,28 @@ namespace mtr
 		void normalize();
 		
 		//gets
-		float getMag() const;
-		float getMagSquared() const;
+		double getMag() const;
+		double getMagSquared() const;
 		Vec4 getUnit() const;
 		Vec4 getProjectionOn(const Vec4 &vec) const;
 		
 		//statics
-		static float angleBetween(const Vec4 &v1, const Vec4 &v2);
+		static double angleBetween(const Vec4 &v1, const Vec4 &v2);
 		
 		//selfOperators
 		const Vec4& operator=(const Vec4 &vec);
 		const Vec4& operator+=(const Vec4 &vec);
 		const Vec4& operator-=(const Vec4 &vec);
-		const Vec4& operator*=(float scalar);
-		const Vec4& operator/=(float scalar);
+		const Vec4& operator*=(double scalar);
+		const Vec4& operator/=(double scalar);
 		const Vec4 operator-() const;
 
 		friend const Vec4 operator+(const Vec4 &v1, const Vec4 &v2);
 		friend const Vec4 operator-(const Vec4 &v1, const Vec4 &v2);
-		friend const float operator*(const Vec4 &v1, const Vec4 &v2);
-		friend const Vec4 operator*(const Vec4 &vec, float scalar);
-		friend const Vec4 operator*(float scalar, const Vec4 &vec);
-		friend const Vec4 operator/(const Vec4 &vec, float scalar);
+		friend const double operator*(const Vec4 &v1, const Vec4 &v2);
+		friend const Vec4 operator*(const Vec4 &vec, double scalar);
+		friend const Vec4 operator*(double scalar, const Vec4 &vec);
+		friend const Vec4 operator/(const Vec4 &vec, double scalar);
 
 		//crossOperators
 		friend const Vec4 operator*(const Mat4 &mat, const Vec4 &vec);
